Add a lint command
Make it easy to use cfn-lint on the templated stacks. The templated
stack is written to a temporary file which cfn-lint is run on. Since
the error messages contain the path to the json node, it's not that
important to preserve the rendered json with matching lines.

Unfortunately there's no special error code for success/failure. We can
just forward the original output.

Having cfn-lint is not a requirement for stack_master and the missing
binary is handled without exploding.