Skip list aliases if all allowed accounts are IDs
When the current account ID doesn't match any of the provided
allowed_accounts, Identity would automatically fetch listed aliases. If
the principal didn't have permissions to retrieve those, this would
result in an error.

When all of the provided "allowed account" values are valid AWS account
IDs, it's currently impossible that an account alias would match, so we
can skip the call, avoiding the potential permissions issue.