Use sorted_set gem instead of relying on it existing in ruby's `set` library
Resolves the following error:

StackMaster::TemplateCompiler::TemplateCompilationFailed Failed to compile a_stack_template.rb
 Caused by: RuntimeError The `SortedSet` class has been extracted from the `set` library. You must use the `sorted_set` gem or other alternatives.

In ruby 3.1, the SortedSet class was removed from the `set` library.

The sparkle_formation gem has resolved this issue, but has yet to release a
new version of the gem.

This commit works around the issue by ensuring the SortedSet class in the
`sorted_set` gem is used by the sparkle_formation gem.